Rename and move Actor class
