+++ src/Digdir.Domain.Dialogporten.Domain/Actors/Actor.cs
@@ -1,7 +1,7 @@
-using Digdir.Domain.Dialogporten.Domain.Dialogs.Entities.Activities;
+using Digdir.Domain.Dialogporten.Domain.Dialogs.Entities;
 using Digdir.Library.Entity.Abstractions;
 
-namespace Digdir.Domain.Dialogporten.Domain.Dialogs.Entities;
+namespace Digdir.Domain.Dialogporten.Domain.Actors;
 
 // **********************************************************************************************************************
 // Rename to Actor
@@ -9,22 +9,13 @@ namespace Digdir.Domain.Dialogporten.Domain.Dialogs.Entities;
 // Created derived types for PerformedBy, SeenLog, Sender(transmission) (and maybe dialog.party / dialog.org)
 // **********************************************************************************************************************
 
-public class DialogActor : IEntity
+public class Actor : IEntity
 {
     public Guid Id { get; set; }
     public DateTimeOffset CreatedAt { get; set; }
     public DateTimeOffset UpdatedAt { get; set; }
     public string? ActorName { get; set; }
     public string? ActorId { get; set; }
-
-    // === Dependent relationships ===
-    // public Guid? DialogId { get; set; }
-    // public DialogEntity Dialog { get; set; } = null!;
-    //
-    // public Guid? ActivityId { get; set; }
-    // public DialogActivity Activity { get; set; } = null!;
-    //
-    // TODO: Add transmissions
 
     public ActorType.Values ActorTypeId { get; set; }
     public ActorType ActorType { get; set; } = null!;
